+++ app/src/main/java/com/edu/xogame/activities/GamePlayActivity.java
@@ -22,24 +22,11 @@ public class GamePlayActivity extends AppCompatActivity {
 
     private Game game;
 
-    private final Handler handler = new Handler();
-    ProgressBar progressBar;
-    int MAX_PROGRESS = 100;
-    int PROGRESS_STEP = 1;
-    int accum = 0;
-    int max_value = 60;
-    Handler handler1 = new Handler();
-    public boolean isRunning;
-    Thread myBackgroundThread;
-    int count=0;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        isRunning = true;
-        progressBar = findViewById(R.id.progressBar);
 
         Intent returnIntent = new Intent();
         setResult(Activity.RESULT_CANCELED, returnIntent);
@@ -73,7 +60,6 @@ public class GamePlayActivity extends AppCompatActivity {
         game = new Game(this, goFirst);
         game.setOpponent(player);
         player.setBoard(game.getBoard());
-        isRunning = true;
         game.start();
     }
 
@@ -92,55 +78,4 @@ public class GamePlayActivity extends AppCompatActivity {
         super.onDestroy();
         MultiPlayerActivity.disconnect(this);
     }
-
-    public void startTime() {
-        if (!isRunning) {
-            return;
-        }
-        accum = 0;
-        progressBar.setMax(MAX_PROGRESS);
-        progressBar.setVisibility(View.VISIBLE);
-        Thread myBackgroundThread = new Thread(backgroundTask, "bgTask");
-        myBackgroundThread.start();
-
-    }
-
-    private Runnable foregroundRunnable = new Runnable() {
-        @Override
-        public void run() {
-            try {
-                progressBar.setProgress((int) ((float) accum / max_value * 100));
-                accum += PROGRESS_STEP;
-
-
-                if (accum >= max_value) {
-                    progressBar.setVisibility(View.INVISIBLE);
-                    isRunning = false;
-                    if(count==0){
-                        game.endGame("Opponent",true);
-                        count++;
-                    }
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-    };
-
-    private Runnable backgroundTask = new Runnable() {
-        @Override
-        public void run() {
-            try {
-                for (int i = 0; i < max_value; i++) {
-                    Thread.sleep(1000);
-
-                    handler1.post(foregroundRunnable);
-
-                }
-
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-    };
 }
